refactor(backend): unify self-management agent tasks (#831)
