+++ backend/app/db/models/__init__.py
@@ -15,7 +15,6 @@ from app.db.models.agent_message_block import AgentMessageBlock
 from app.db.models.auth_audit_event import AuthAuditEvent
 from app.db.models.auth_legacy_refresh_revocation import AuthLegacyRefreshRevocation
 from app.db.models.auth_refresh_session import AuthRefreshSession
-from app.db.models.built_in_follow_up_task import BuiltInFollowUpTask
 from app.db.models.conversation_thread import ConversationThread
 from app.db.models.external_session_directory_cache import (
     ExternalSessionDirectoryCacheEntry,
@@ -23,7 +22,7 @@ from app.db.models.external_session_directory_cache import (
 from app.db.models.hub_a2a_agent_allowlist import HubA2AAgentAllowlistEntry
 from app.db.models.hub_a2a_user_credential import HubA2AUserCredential
 from app.db.models.invitation import Invitation, InvitationStatus
-from app.db.models.self_management_dispatch_task import SelfManagementDispatchTask
+from app.db.models.self_management_agent_task import SelfManagementAgentTask
 from app.db.models.shortcut import Shortcut
 from app.db.models.user import User
 from app.db.models.user_agent_availability_snapshot import UserAgentAvailabilitySnapshot
@@ -40,14 +39,13 @@ __all__ = [
     "AuthAuditEvent",
     "AuthLegacyRefreshRevocation",
     "AuthRefreshSession",
-    "BuiltInFollowUpTask",
     "ConversationThread",
     "HubA2AAgentAllowlistEntry",
     "HubA2AUserCredential",
     "Invitation",
     "InvitationStatus",
     "ExternalSessionDirectoryCacheEntry",
-    "SelfManagementDispatchTask",
+    "SelfManagementAgentTask",
     "User",
     "Shortcut",
     "UserAgentAvailabilitySnapshot",
